+++ project/app/src/main/java/com/skowronsky/snkrs/database/Favorite.java
@@ -9,15 +9,15 @@ import com.google.gson.annotations.SerializedName;
 @Entity
 public class Favorite {
     @PrimaryKey(autoGenerate = true)
-    public int idFavoriteShoes;
-    public int idShoes;
-    public int idBase;
+    public long idFavoriteShoes;
+    public long idShoes;
+    public long idBase;
     public double size;
 
     public Favorite(){};
 
     @Ignore
-    public Favorite(Shoes shoes, double size, int idBase){
+    public Favorite(Shoes shoes, double size, long idBase){
         this.idShoes = shoes.idShoes;
         this.size = size;
         this.idBase = idBase;
